+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+### Added
+- all files: added JavaDoc
+- `ImmutableCollection`, `ImmutableList`, `ImmutableMap`, `ImmutableSet`: added length checks
+
 ## [1.0] - 2021-07-09
 ### Added
 - all files: initial implementations